Bug fixes to ensure all tests are satisfied
Code for mutliple modules adjusted to ensure that all tests in the tests
folder complete successfully when run under Python 3.6. Most fixes had
to do with the fact that Python 2 and 3 cast integers differently.
However, one fix had to do with a new warning added to the method for
choosing untrans and trans curves.

Note that test_robustness.py was moved out of this repository.

 On branch develop
 Changes to be committed:
	modified:   LICENSE.txt
	modified:   eugene/src/data_prep/initial_conditions.py
	modified:   eugene/src/data_prep/metaparameters.py
	modified:   eugene/src/dd_scan.py
	modified:   eugene/tests/test_initial_conditions.py
	deleted:    eugene/tests/test_robustness.py

 Untracked files:
	pyproject.toml
	setup.cfg

+++ eugene/src/data_prep/metaparameters.py
@@ -91,13 +91,13 @@ def tune_ic_selection(
 
     # set up ranges
     if max_len is None:
-        max_len = 2 * min_len
+        max_len = int(2 * min_len)
     if num_frags_range is None:
         lo = int(min_series_len / max_len) 
         hi = int(min_series_len /  min_len)  
         num_frags_range = [lo, hi]
     if max_reps  is None:
-        max_reps = num_frags_range[1] / 10
+        max_reps = int(num_frags_range[1] / 10)
     print("num_frags_range = {}".format(num_frags_range))
     print("min reps = {}".format(min_reps))
     print("max reps = {}".format(max_reps))
